# contracts/api.yaml
openapi: "3.1.0"
info:
  title: Asphalt Ingestion and Query API
  version: "1.0.0"
  description: |
    REST API for the Asphalt road anomaly detection platform.

    Two surface areas:
    - Ingestion: accepts event batches from the Android SDK
    - Query: serves aggregated road quality data for map rendering

    All coordinates use WGS84. All timestamps are Unix milliseconds UTC.
servers:
  - url: https://api.asphalt.io/v1
    description: Production
paths:
...
  /map/clusters:
    get:
      operationId: getClusters
      summary: Get anomaly clusters within a bounding box
      parameters:
        - name: min_lat
          in: query
          required: true
          schema:
            type: number
        - name: min_lon
          in: query
          required: true
          schema:
            type: number
        - name: max_lat
          in: query
          required: true
          schema:
            type: number
        - name: max_lon
          in: query
          required: true
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
            maximum: 1000
      responses:
        "200":
          description: List of clusters
          content:
            application/json:
              schema:
                type: object
                properties:
                  clusters:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnomalyCluster"
...
components:
  schemas:
...
    AnomalyCluster:
      type: object
      required: [cluster_id, latitude, longitude, anomaly_type, confidence, event_count, avg_intensity, last_seen_ms]
      properties:
        cluster_id:
          type: string
          format: uuid
        latitude:
          type: number
        longitude:
          type: number
        anomaly_type:
          type: string
          enum: [pothole, bump, rough_patch, unknown]
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Confidence that this cluster represents a real anomaly, based on event count, consistency of reports, and recency.
        event_count:
          type: integer
          description: Total number of raw events that contributed to this cluster
        avg_intensity:
          type: number
          minimum: 0
          maximum: 1
        last_seen_ms:
          type: integer
          description: Timestamp of most recent contributing event
        radius_m:
          type: number
          description: Approximate radius of the cluster in metres
...